+++ src/c++/lib/applications/starling/starling_pos_processor.cpp
@@ -677,9 +677,9 @@ process_pos_snp_digt(
     }
 
     double homRefLogProb(0);
+    ActiveRegionId activeRegionId(getActiveRegionDetector().getActiveRegionId(pos));
     for (unsigned sampleIndex(0); sampleIndex < sampleCount; ++sampleIndex)
     {
-        ActiveRegionId activeRegionId(getActiveRegionDetector(sampleIndex).getActiveRegionId(pos));
         updateSnvLocusWithSampleInfo(
             _opt, sample(sampleIndex), callerPloidy[sampleIndex], groupLocusPloidy[sampleIndex],
             allDgt[sampleIndex], sampleIndex, activeRegionId, getCandidateSnvBuffer(), *locusPtr, homRefLogProb);
@@ -1716,10 +1716,10 @@ process_pos_indel_digt(const pos_t pos)
 
             // Add sample-dependent info to locus:
             double homRefLogProb(0);
+            ActiveRegionId activeRegionId(getActiveRegionDetector().getActiveRegionId(pos));
             for (unsigned sampleIndex(0); sampleIndex < sampleCount; ++sampleIndex)
             {
                 auto& sampleInfo(sample(sampleIndex));
-                ActiveRegionId activeRegionId(getActiveRegionDetector(sampleIndex).getActiveRegionId(pos));
                 updateIndelLocusWithSampleInfo(
                     _opt, _dopt, topVariantAlleleGroup, topVariantAlleleIndexPerSample[sampleIndex], emptyGroup,
                     sampleInfo.sampleOptions,
@@ -1871,10 +1871,10 @@ process_pos_indel_digt(const pos_t pos)
 
             // add sample-dependent info:
             double homRefLogProb(0);
+            ActiveRegionId activeRegionId(getActiveRegionDetector().getActiveRegionId(pos));
             for (unsigned sampleIndex(0); sampleIndex < sampleCount; ++sampleIndex)
             {
                 auto& sif(sample(sampleIndex));
-                ActiveRegionId activeRegionId(getActiveRegionDetector(sampleIndex).getActiveRegionId(pos));
                 updateIndelLocusWithSampleInfo(
                     _opt, _dopt, fakeForcedOutputAlleleGroup, fakeTopVariantAlleleIndexPerSample, topVariantAlleleGroup,
                     sif.sampleOptions, callerPloidy[sampleIndex], groupLocusPloidy[sampleIndex], sampleIndex, activeRegionId,